+++ lib/jsapi/dsl/definitions.rb
@@ -5,6 +5,17 @@ module Jsapi
     # Used to define top-level API components.
     class Definitions < Node
 
+      # Specifies the general default values for +type+.
+      #
+      #   default 'array', read: [], write: []
+      #
+      def default(type, **keywords, &block)
+        _define('default', type.inspect) do
+          default = _meta_model.add_default(type, keywords)
+          Node.new(default, &block) if block
+        end
+      end
+
       # Includes API definitions from +klasses+.
       def include(*klasses)
         klasses.each do |klass|